// YiPet/src/definitions.d.ts
export interface ExtensionData {
    isEnabled: boolean;
    isReady: boolean;
    isAllowedFileSchemeAccess: boolean;
    settings: UserSettings;
    news: News[];
    shortcuts: Shortcuts;
    colorScheme: ParsedColorSchemeConfig;
    forcedScheme: 'dark' | 'light' | null;
    activeTab: TabInfo;
    uiHighlights: string[];
}

export type DevFixType = 'detector' | 'dynamic' | 'filter' | 'static';

export type DevToolsData = Record<DevFixType, string>;

export interface TabData {
    type: MessageTypeBGtoCS;
    data?: any;
}
// ...
export interface Theme {
    mode: FilterMode;
    brightness: number;
    contrast: number;
    grayscale: number;
    sepia: number;
    useFont: boolean;
    fontFamily: string;
    textStroke: number;
    engine: ThemeEngine;
    stylesheet: string;
    darkSchemeBackgroundColor: string;
    darkSchemeTextColor: string;
    lightSchemeBackgroundColor: string;
    lightSchemeTextColor: string;
    scrollbarColor: '' | 'auto' | string;
    selectionColor: '' | 'auto' | string;
    styleSystemControls: boolean;
    lightColorScheme: string;
    darkColorScheme: string;
    immediateModify: boolean;
}

export interface CustomSiteConfig {
    url: string[];
    theme: Theme;
    builtIn?: boolean;
}

export interface ThemePreset {
    id: string;
    name: string;
    urls: string[];
    theme: Theme;
}

export interface UserSettings {
    schemeVersion: number;
    enabled: boolean;
    fetchNews: boolean;
    theme: Theme;
    presets: ThemePreset[];
    customThemes: CustomSiteConfig[];
    enabledByDefault: boolean;
    enabledFor: string[];
    disabledFor: string[];
    changeBrowserTheme: boolean;
    syncSettings: boolean;
    syncSitesFixes: boolean;
    previewNewDesign: boolean;
    previewNewestDesign: boolean;
    enableForPDF: boolean;
    enableForProtectedPages: boolean;
    enableContextMenus: boolean;
    detectDarkTheme: boolean;
}

export interface TabInfo {
    url: string;
    id: number | null;
    documentId: string | null;
    isProtected: boolean;
    isInjected: boolean | null;
    isInDarkList: boolean;
    isDarkThemeDetected: boolean | null;
}
// ...
export interface Shortcuts {
    [name: string]: string;
}
// ...
export interface News {
    id: string;
    date: string;
    url: string;
    headline: string;
    read?: boolean;
    displayed?: boolean;
    badge?: string;
    icon?: string;
}

// These values need to match those in Manifest
export type Command = 'toggle' | 'addSite' | 'switchEngine';